examples/pipelines/gpt_4o_multimodal_rag/requirements.txt
pathway[xpack-llm]
openparse==0.5.6
python-dotenv==1.0.1
unstructured[all-docs]==0.10.28
mpmath==1.3.0
